move data preprocessing to contrib
