+++ pyro/contrib/examples/scanvi_data.py
